feat: add file schema and update file repository

+++ services/fileService.ts
@@ -1,13 +1,26 @@
 import { IfileRepository, IresponseUploadFile } from "@/utils/interface";
 
-export class FileService implements IfileRepository {
+export class FileService {
   constructor(private fileRepository: IfileRepository) {}
 
-  async uploadFile(file: File, prefix?: string): Promise<IresponseUploadFile> {
-    return this.fileRepository.uploadFile(file, prefix);
+  async uploadFile(
+    userID: string,
+    file: File,
+    prefix?: string
+  ): Promise<IresponseUploadFile> {
+    return this.fileRepository.uploadFile(userID, file, prefix);
   }
 
-  async deleteFile(key: string): Promise<void> {
+  async deleteFile(userID: string, key: string): Promise<void> {
+    const file = await this.fileRepository.getByURL(key);
+    if (!file) {
+      throw new Error("File not found");
+    }
+
+    if (file.userID !== userID) {
+      throw new Error("Unauthorized to delete this file");
+    }
+
     return this.fileRepository.deleteFile(key);
   }
 }
